Bug 1821473 [wpt PR 38919] - Fix failure to include cookies via Storage Access API on navigations, a=testonly
Automatic update from web-platform-tests
Fix failure to include cookies via Storage Access API on navigations

This fix is not quite ideal, since it relies on the existing
has_storage_access bool in CommonNavigationParams, which is only set
to true under strict circumstances. A proper fix will introduce a new
bool in BeginNavigationParams which will be true in more situations,
and will use that instead in NavigationURLLoaderImpl.

However, given that branch cut is around the corner, and I'm OOO all
next week, and this hack will fix the overwhelmingly-common case
(frame reloads), this is ok for now.

Bug: 1423092
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4326169
Commit-Queue: Chris Fredrickson <cfredricchromium.org>
Cr-Commit-Position: refs/heads/main{#1115532}

--

wpt-commits: 1b710fdad025467914012012dd96b5b72fcd27b4
wpt-pr: 38919

UltraBlame original commit: 18107de5722a0958f8f569d2cc3ae1dfe36fe976

+++ testing/web-platform/tests/storage-access-api/helpers.js
@@ -130,7 +130,7 @@ async function DeleteCookieInFrame(frame, name, params) {
 
 
 
-async function CanFrameWriteCookies(frame) {
+async function CanFrameWriteCookies(frame, keep_after_writing = false) {
   const cookie_suffix = "Secure;SameSite=None;Path=/";
   await DeleteCookieInFrame(frame, "cookie", cookie_suffix);
   await DeleteCookieInFrame(frame, "foo", cookie_suffix);
@@ -142,8 +142,10 @@ async function CanFrameWriteCookies(frame) {
   const can_write = cookieStringHasCookie("cookie", "monster", cookies) &&
       cookieStringHasCookie("foo", "bar", cookies);
 
-  await DeleteCookieInFrame(frame, "cookie", cookie_suffix);
-  await DeleteCookieInFrame(frame, "foo", cookie_suffix);
+  if (!keep_after_writing) {
+    await DeleteCookieInFrame(frame, "cookie", cookie_suffix);
+    await DeleteCookieInFrame(frame, "foo", cookie_suffix);
+  }
 
   return can_write;
 }
